+++ doc/changelog.md
@@ -4,7 +4,8 @@
   - More robust response stream parsing (#436)
   - Add `docker.dockerFileDir` and `docker.dockerFile` to the properties configuration provider. (#438)
   - Fix splitting of bind volumes for Windows pathes (#443)
-    
+  - Add new build config option `user` for switching the user at the end of the Dockerfile. `docker.user` can be used
+    for the properties configuration provider (#441)    
 * **0.15.1** (2016-05-03)
   - Fix push / pull progress bar ([#91](https://github.com/fabric8io/docker-maven-plugin/issues/91))
   - Allow empty environment variable ([#434](https://github.com/fabric8io/docker-maven-plugin/issues/434))